feat: Crear el controlador de usuario. #45
